backend/requirements.txt
fastapi
uvicorn
retailcrm==5.3.0
supabase
python-dotenv
pydantic
pydantic-settings
httpx
loguru
slowapi
python-multipart
